requirements.txt
python-telegram-bot[job-queue]==21.5
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.1
